Feat: added user count api
