+++ src/app/api/manager/route.ts
@@ -33,3 +33,5 @@ export async function GET(_request: NextRequest) {
 
 }
 
+
+
